+++ packages/project/src/parse/from-app-state.ts
@@ -2,19 +2,21 @@
 
 import * as l from '@openfn/lexicon';
 import { Provisioner } from '@openfn/lexicon/lightning';
-import { OpenfnConfig, Project } from '../Project';
+import { Project } from '../Project';
 import { yamlToJson } from '../util/yaml';
 import renameKeys from '../util/rename-keys';
+import { WorkspaceConfig } from '../util/config';
 
 // Extra metadata used to init the project
-type FromAppStateConfig = {
+export type FromAppStateConfig = {
   endpoint: string;
   env?: string;
   fetchedAt?: string;
   format?: 'json' | 'yaml';
 
   // Allow workspace config to be passed
-  repo: OpenfnConfig;
+  // TODO can we just pass a Workspace?
+  config: WorkspaceConfig;
 };
 
 function slugify(text) {
@@ -52,6 +54,7 @@ export default (state: Provisioner.Project, config: FromAppStateConfig) => {
 
   proj.openfn = {
     uuid: id,
+    name: name,
     endpoint: config.endpoint,
     env: config.env,
     inserted_at,
@@ -65,7 +68,7 @@ export default (state: Provisioner.Project, config: FromAppStateConfig) => {
 
   proj.workflows = state.workflows.map(mapWorkflow);
 
-  return new Project(proj as l.Project, config?.repo);
+  return new Project(proj as l.Project, config?.config);
 };
 
 const mapTriggerEdgeCondition = (edge: Provisioner.Edge) => {