+++ packages/project/src/serialize/to-app-state.ts
@@ -26,7 +26,7 @@ export default function (project: Project, options: Options = {}) {
 
   const shouldReturnYaml =
     options.format === 'yaml' ||
-    (!options.format && project.repo.formats.project === 'yaml');
+    (!options.format && project.config.formats.project === 'yaml');
 
   if (shouldReturnYaml) {
     return jsonToYaml(state);
@@ -94,6 +94,7 @@ const mapWorkflow = (workflow) => {
         id: rules.openfn?.uuid ?? randomUUID(),
         target_job_id: lookup[next],
         enabled: !rules.disabled,
+        source_trigger_id: null, // lightning complains if this isn't set, even if its falsy :(
       };
 
       if (isTrigger) {