+++ packages/project/test/parse/from-app-state.test.ts
@@ -75,6 +75,7 @@ test('should create a Project from prov state with app project metadata', (t) =>
   t.deepEqual(project.openfn, {
     env: 'test',
     uuid: state.id,
+    name: 'My Workflow',
     endpoint: config.endpoint,
     inserted_at: state.inserted_at,
     updated_at: state.updated_at,
@@ -114,6 +115,7 @@ test('should create a Project from prov state with a workflow', (t) => {
   t.deepEqual(project.workflows[0].toJSON(), {
     id: 'my-workflow',
     name: 'My Workflow',
+    history: [],
     steps: [
       {
         id: 'trigger',